requirements.txt
langchain
langchain-groq
langchain-community
langchain-huggingface
langgraph
chromadb
openpyxl
pandas
streamlit
python-dotenv
sentence-transformers
fastapi